+++ src/Translator.hs
@@ -3,34 +3,34 @@ module Translator
   )
   where
 
-import Dictionary (searchTranslation, sortOutput)
+import Dictionary (Answer, searchTranslation, sortOutput)
 import Effects.Console (Console, cancelInput, closeInput, exitSuccess, getHistory, getInput,
                         initializeInput)
 import Effects.PrettyPrint (Line (NewLine), PrettyPrint, pprint, putColorDoc)
 import Env (Env)
-import Parse (fromTibetScript, fromWylieScript, parseTibetanInput, parseWylieInput, toTibetan,
-              toWylie)
+import Parse (fromTibetScript, fromWylieScript, parseEither, parseTibetanInput, parseWylieInput,
+              tibetanWord, toTibetan, toWylie, wylieWord)
 import Pretty (blue, red, viewTranslations, withHeaderSpaces, yellow)
 import Type (HibetError (..))
+import Utility (showT)
 
-import Control.Monad.Except (Except, forever, runExcept)
+import Control.Monad.Except (Except, forever, liftEither, runExcept)
 import Control.Parallel.Strategies (parList, rseq, using)
 import Data.List (foldl')
 import Data.Maybe (mapMaybe)
 import Data.Text (Text)
 import qualified Data.Text as T
 import Polysemy (Members, Sem)
-import Polysemy.Error (Error, fromEither)
+import Polysemy.Error (Error)
 import Polysemy.Reader (Reader, ask)
 import Polysemy.Resource (Resource, bracketOnError)
+import Polysemy.Trace (Trace, trace)
 import Prettyprinter (Doc)
 import Prettyprinter.Render.Terminal (AnsiStyle)
 import System.Console.Haskeline.History (History, historyLines)
 import System.Console.Haskeline.IO (InputState)
 
-import Polysemy.Trace (Trace)
--- import qualified Debug.Trace as Debug
--- import Polysemy.Trace (trace)
+import qualified Debug.Trace as Debug
 
 -- | Load environment and start loop dialog
 translator :: Members [PrettyPrint, Trace, Resource, Console, Error HibetError, Reader Env ] r
@@ -38,16 +38,23 @@ translator :: Members [PrettyPrint, Trace, Resource, Console, Error HibetError, 
 translator = bracketOnError
   initializeInput
   cancelInput -- This will only be called if an exception such as a SigINT is received.
-  (\inputState -> loopDialog inputState >> closeInput inputState)
-  -- (\inputState -> runReaderT (testDialog inputState) env >> closeInput inputState) -- for tests
+  $ \inputState -> do
+      putColorDoc blue NewLine "Please, input a word with tibetan script or wylie transcription!"
+      loopDialog inputState
+      closeInput inputState
 
 
 -- Looped dialog with user
-loopDialog :: Members [PrettyPrint, Trace, Console, Error HibetError, Reader Env ] r
+loopDialog :: Members
+  [ PrettyPrint
+  , Trace
+  , Console
+  , Error HibetError
+  , Reader Env ] r
   => InputState
   -> Sem r ()
 loopDialog inputState = forever $ do
-    putColorDoc blue NewLine "Input a tibetan word in wylie transcription, please."
+    putColorDoc blue NewLine "Your request:"
     mQuery <- getInput inputState "> "
     case T.strip . T.pack <$> mQuery of
         Nothing -> pure ()
@@ -57,46 +64,76 @@ loopDialog inputState = forever $ do
         Just ":h" -> do
             history <- fromHistory <$> getHistory inputState
             mapM_ (putColorDoc id NewLine) history
-        Just query -> do
+        Just input -> do
             env :: Env <- ask
-            (answer, isEmpty) <- fromEither $ runExcept $ getAnswer query env
-            if isEmpty
-              then putColorDoc red NewLine "Nothing found"
-              else pprint answer
+            case runExcept $ getAnswer input env of
+              Left err -> do
+                trace $ show err
+                putColorDoc red NewLine "Nothing found"
+              Right (query, answer) -> if null answer
+                then putColorDoc red NewLine "Nothing found"
+                else pprint $ mkOutput query answer
+
+
+getAnswer :: Text -> Env -> Except HibetError (Text, [Answer])
+getAnswer query env = do
+  -- 1. Detect script of input
+  script <- detectScript query
+  -- Debug.traceM $ show script
+  -- 2. Prepare wylie and tibetan scripts
+  (queryWylie, queryTibet) <- boilQuery script query env
+  -- Debug.traceM ("queryTibet " <> T.unpack queryTibet)
+  -- 3. Translate
+  let dscValues = mapMaybe (searchTranslation queryWylie) env.dictionaryMeta `using` parList rseq
+  -- 5. Return
+  pure (queryTibet, dscValues)
+
 
 fromHistory :: History -> [Text]
 fromHistory = foldl' (\ a x -> T.pack x : a) [] . filter (/=":h") . historyLines
 
+data Script = T | W
+  deriving stock Show
 
-getAnswer :: Text -> Env -> Except HibetError (Doc AnsiStyle, Bool)
-getAnswer query env = do
-  let toWylie' = toWylie env.tibetWylieMap . parseTibetanInput env.radixTibet
-      -- 1. Parse text to tibetan script,
-      -- 2. check tibetan script is valid,
-      -- 3. convert to Wylie.
-      queryWylie = case runExcept $ toWylie' query  of
-        Left _      -> query
-        Right wylie -> if null wylie then query else T.intercalate " " $ map fromWylieScript wylie
-      dscValues = mapMaybe (searchTranslation queryWylie) env.dictionaryMeta `using` parList rseq
-  let list = sortOutput dscValues
-  let toTibetan' = toTibetan env.wylieTibetMap . parseWylieInput env.radixWylie
-  -- list <- traverse (separator [37] toTibetan') dictMeta
-  let (translations, isEmpty) = (viewTranslations list, list == mempty)
-  query' <- if query == queryWylie
-    then do
-      tibetScript <- toTibetan' queryWylie
-      pure $ T.intercalate "་" $ fromTibetScript <$> tibetScript
-    else pure query
-  pure (withHeaderSpaces yellow query' translations, isEmpty)
+detectScript :: Text -> Except HibetError Script
+detectScript query = do
+  let eWylie = parseEither wylieWord query
+  let eTibetan = parseEither tibetanWord query
+  liftEither $ case (eWylie, eTibetan) of
+    (Left _, Right _)  -> Right T -- likely tibetan
+    (Right _, _) -> Right W -- likely wylie
+    (Left ew, Left et)  -> do
+      Debug.traceM ("Left ew: " <> show ew)
+      Debug.traceM ("Left et: " <> show et)
+      Left $ UnknownError (showT et <> "\n" <> showT ew)
 
+boilQuery :: Script -> Text -> Env -> Except HibetError (Text, Text)
+boilQuery script query env = liftEither $ case script of
+  W -> do
+    wList <- parseWylieInput env.radixWylie query
+    -- Debug.traceM ("wList " <> show wList)
+    wylieText <- fromWylieScript wList
+    -- Debug.traceM ("wylieText " <> show wylieText)
+    tList <- toTibetan env.wylieTibetMap wList
+    -- Debug.traceM ("tList " <> show tList)
+    tibetanText <- fromTibetScript tList
+    -- Debug.traceM ("tibetanText " <> show tibetanText)
+    Right (wylieText, tibetanText)
+  T -> do
+    -- 2.1. Parse text to tibetan script,
+    -- 2.2. check tibetan script is valid,
+    -- 2.3. convert to Wylie.
+    tList <- parseTibetanInput env.radixTibet query
+    tibetanText <- fromTibetScript tList
+    -- Debug.traceM ("tList " <> show tList)
+    wList <- toWylie env.tibetWylieMap tList
+    -- Debug.traceM ("wList " <> show wList)
+    wylieText <- fromWylieScript wList
+    -- Debug.traceM ("wylieText " <> show wylieText)
+    Right (wylieText, tibetanText)
 
--- testDialog :: InputState -> Hibet ()
--- testDialog inputState = ReaderT $ \env -> forever $ do
---   putColorDoc blue NewLine "Which a tibetan word to translate?"
---   let answerE = runExcept $ getAnswer "mo" env
---   case answerE of
---       Left err -> putColorDoc red NewLine $ T.pack $ ME.errorBundlePretty err
---       Right (answer, isEmpty) ->
---           if isEmpty then putColorDoc red NewLine "Nothing found"
---           else print answer
---   exitSuccess
+mkOutput :: Text -> [Answer] -> Doc AnsiStyle
+mkOutput query answers
+  = withHeaderSpaces yellow query
+  $ viewTranslations
+  $ sortOutput answers